RDKB-36286 - Changes for segmentation fault observed with 2 consumers
Reason for change: When 2 consumers are running parallel, there is possibility of add/remove row getting failed. As the consumer running in parallel are identical. Since the add/remove row is getting failed which is called inside a macro. The macro is written such that it will exit the function without unsubscribing the events subscribed by the function. This result in possibility of the consumer 1 and 2 which are running in parallel to crash.
For example: Consider the consumer running in parallel as 1 and 2. When the consumer 1 is failed in add/remove row, then while the consumer 1 is returning from testTableEvents()->testEvents()->testconsumer(). There is possibility of the consumer 2 is adding the same data which consumer 1 is subscribed for and the consumer 1 is returning without unsubscribing. So, consumer 1 also gets the event and the event handler is called, but the data are not there in consumer list as the consumer is written such that to delete all the list content before calling add/remove row every time.
Test Procedure: Test and verified with ubuntu VM
Risks: Low
Signed-off-by: Gururaaja E S R <gururaja_erodesriranganramlingham@comcast.com>

(cherry picked from commit 5aa70fa7a954aa9034385f682bb909d21b86987c)
Change-Id: I39fdd6da783ddaf146a41012aca5ad1896fa7711